refactor: Move coordinate transformation tests to separate file
